refactor: 상세보기 뷰 롤백
