+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         android:value="false" />
 
     <application
-        android:name="com.team.recordream.application.RecordreamApp"
+        android:name=".application.RecordreamApp"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -39,43 +39,42 @@
         android:theme="@style/Theme.RecorDream"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
+        <activity
+            android:name=".presentation.detail.DetailActivity"
+            android:exported="false" />
 
         <meta-data
             android:name="com.facebook.sdk.ApplicationId"
             android:value="@string/facebook_app_id" />
 
         <activity
-            android:name="com.team.recordream.presentation.search.SearchActivity"
+            android:name=".presentation.search.SearchActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan" />
-
         <activity
-            android:name="com.team.recordream.presentation.record.RecordActivity"
+            android:name=".presentation.record.RecordActivity"
             android:exported="true"
             android:screenOrientation="portrait" />
         <activity
-            android:name="com.team.recordream.presentation.MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name="com.team.recordream.presentation.login.LoginActivity"
+            android:name=".presentation.login.LoginActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name="com.team.recordream.presentation.mypage.MypageActivity"
+            android:name=".presentation.mypage.MypageActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
-
         <activity
-            android:name="com.team.recordream.presentation.splash.SplashActivity"
+            android:name=".presentation.splash.SplashActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/SplashTheme">
-
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
@@ -85,14 +84,11 @@
             android:exported="true"
             android:launchMode="singleTask"
             android:screenOrientation="portrait">
-
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
 
-
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
-
 
                 <!-- Redirect URI: "kakao{NATIVE_APP_KEY}://oauth" -->
                 <!-- scheme-example: "kakao{NATIVE_APP_KEY}" -->
@@ -100,9 +96,7 @@
                     android:host="oauth"
                     android:scheme="@string/appKey_manifest_kakao" />
             </intent-filter>
-        </activity>
-
-        <!-- 외부 저장소 사용을 위한 FileProvider auth 추가. -->
+        </activity> <!-- 외부 저장소 사용을 위한 FileProvider auth 추가. -->
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.fileprovider"
@@ -110,13 +104,11 @@
             android:grantUriPermissions="true">
             <meta-data
                 android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/path">
-
-            </meta-data>
+                android:resource="@xml/path"></meta-data>
         </provider>
 
         <service
-            android:name="com.team.recordream.util.RecorDreamFireBaseMessagingService"
+            android:name=".util.RecorDreamFireBaseMessagingService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
