+++ app/src/main/java/com/team/recordream/presentation/detail/adapter/ContentAdapter.kt
@@ -1,11 +1,11 @@
 package com.team.recordream.presentation.detail.adapter
 
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.team.recordream.presentation.detail.DetailBottomSheetFragment
 
-class ContentAdapter(detailFragment: DetailBottomSheetFragment) :
-    FragmentStateAdapter(detailFragment) {
+class ContentAdapter(activity: FragmentActivity) :
+    FragmentStateAdapter(activity) {
     val fragments: MutableList<Fragment> = mutableListOf()
 
     override fun getItemCount(): Int = fragments.size
